- Deleted TableFilterState   - It is just an extention of method for the JTableFilter   - It just stored the filter items in an array which can be done right from the JTableFilter without the need of another class.
